refactor(kill-matrix-table): move pivot logic into KillMatrixFormatter, table receives pre-computed data
KillMatrixFormatter.pivot() extracts killers × victims matrix from KillMatrixViewRow[].
KillMatrixTable now accepts KillMatrixPivotData instead of raw rows and only handles
column construction + rendering — no data transformation useMemo.
match-stats and series-stats call KillMatrixFormatter.pivot() before passing to the table.
